remove the explicit setReadTimeout in configureHook
This is board-specific (EasySYNC for instance uses a serial protocol
and requires a nonzero timeout). The current implementation of
updateHook should anyway ensure that it does not block, regardless
of the timeout, given that it only reads the packets that have been
reported as being available by the board.

If some CAN drivers work fine without blocking, they should set
the timeout to zero in their implementation.